app/src/main/java/com/selva/androidproficiencyexercise/model/CountryData.java

package com.selva.androidproficiencyexercise.model;

import java.util.ArrayList;
import java.util.List;
import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

public class CountryData {

    @SerializedName("title")
    @Expose
    private String title;
    @SerializedName("rows")
    @Expose
    private List<Row> rows = new ArrayList<>();

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public CountryData withTitle(String title) {
        this.title = title;
        return this;
    }

    public List<Row> getRows() {
        return rows;
    }

    public void setRows(List<Row> rows) {
        this.rows = rows;
    }

    public CountryData withRows(List<Row> rows) {
        this.rows = rows;
        return this;
    }

}
